# contracts/openapi/bsmart-v1.yaml
openapi: 3.1.0
info:
  title: bSmart Client API
  version: 1.8.0
  description: Versioned research and legacy account contract. Current native Google/Apple login uses Supabase Auth directly; wallet registry is specified in supabase-wallet.yaml and docs/contracts/supabase_account.md.
servers:
  - url: https://api.bsmart.today
security:
  - installationBearer: []
paths:
...
  /v1/trade-feed:
    get:
      operationId: getTradeFeed
      description: No-store. Verified opinion-attributed orders with explicit public amount-sharing consent. Source-off returns 503. Partial fills aggregate per order; live pagination.
      parameters:
        - { in: query, name: offset, schema: { type: integer, minimum: 0, default: 0 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 50, default: 30 } }
        - { in: query, name: profileId, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Newest-first public executions; no wallet or account identifiers.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TradeFeedPage' }
        "401": { description: Installation authentication required. }
        "503": { description: Verified Feed unavailable; not zero trades. }
  /v1/public-traders/{profileId}:
    get:
      operationId: getPublicTrader
      description: No-store. Rechecks current public Feed consent.
      parameters:
        - { in: path, name: profileId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Current public profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeedPublicProfile' }
        "401": { description: Installation authentication required. }
        "404": { description: Unknown or private profile. }
        "503": { description: Public profiles unavailable. }
  /v1/opinions/{opinionId}/traders:
    get:
      operationId: getOpinionTraders
      description: Distinct real traders attributed before order submission. No-store. Private users count but are not listed. Disabled until independently verified real execution is connected.
      parameters:
        - { in: path, name: opinionId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: offset, schema: { type: integer, minimum: 0, default: 0 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 50, default: 30 } }
      responses:
        "200":
          description: Verified count and consented public profiles ordered by latest fill descending.
          content:
            application/json:
              schema:
                type: object
                required: [totalTraders, publicTraders, traders, nextOffset]
                properties:
                  totalTraders: { type: integer, minimum: 0 }
                  publicTraders: { type: integer, minimum: 0 }
                  nextOffset: { type: [integer, 'null'], minimum: 0 }
                  traders:
                    type: array
                    items:
                      type: object
                      required: [id, nickname, avatarURL, side, tradedAt]
                      properties:
                        id: { type: string, format: uuid }
                        nickname: { type: string, maxLength: 28 }
                        avatarURL: { type: [string, 'null'], format: uri }
                        side: { type: string, enum: [long, short] }
                        tradedAt: { type: string, format: date-time }
        "401": { description: Installation authentication required. }
        "503": { description: Real execution statistics unavailable; do not display zero. }
...
components:
...
  securitySchemes:
    accountBearer:
      type: http
      scheme: bearer
      description: Opaque verified account session, distinct from installation authentication.
    installationBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-installation-token
  schemas:
    FeedPublicProfile:
      type: object
      additionalProperties: false
      required: [id, nickname, avatarURL]
      properties:
        id: { type: string, format: uuid }
        nickname: { type: string, minLength: 1, maxLength: 28 }
        avatarURL: { type: [string, 'null'], format: uri }
    TradeFeedPage:
      type: object
      additionalProperties: false
      required: [items, nextOffset]
      properties:
        nextOffset: { type: [integer, 'null'], minimum: 0 }
        items:
          type: array
          maxItems: 50
          items: { $ref: '#/components/schemas/TradeFeedItem' }
    TradeFeedItem:
      type: object
      additionalProperties: false
      required: [id, trader, opinion, side, notionalUSD, marketCoin, executedAt]
      properties:
        id: { type: string, format: uuid }
        trader: { $ref: '#/components/schemas/FeedPublicProfile' }
        opinion: { $ref: '#/components/schemas/SmartAccountUpdate' }
        side: { type: string, enum: [long, short] }
        notionalUSD: { type: string, maxLength: 64, pattern: '^[0-9]+(\.[0-9]+)?$' }
        marketCoin: { type: string, minLength: 1, maxLength: 64 }
        executedAt: { type: string, format: date-time }
...
    PriceCandle:
      type: object
      required: [day, open, high, low, close, volume]
      properties:
        day: { type: string, format: date }
        open: { type: number }
        high: { type: number }
        low: { type: number }
        close: { type: number }
        volume: { type: integer }
    SmartAccountPriceEvidence:
      type: object
      description: Real daily OHLC context around one time-stamped Smart Account view; it is audit evidence, not proof of causality.
      required: [ticker, viewDay, viewPrice, latestDay, latestPrice, source, candles]
      properties:
        ticker: { type: string }
        viewDay: { type: string, format: date }
        viewPrice: { type: number }
        latestDay: { type: string, format: date }
        latestPrice: { type: number }
        responsePercent: { type: [number, "null"] }
        source: { type: string }
        candles:
          type: array
          items:
            $ref: "#/components/schemas/PriceCandle"
        opinionMarkers:
          type: array
          description: Up to ten Score-positive views; the earliest positive anchor first, then the highest-contributing remaining views.
          items:
            $ref: "#/components/schemas/SmartAccountOpinionMarker"
    SmartAccountOpinionMarker:
      type: object
      required: [id, publishedAt, viewDay, viewPrice, direction, contribution, horizon, thesis]
      properties:
        id: { type: string, format: uuid }
        publishedAt: { type: string, format: date-time }
        viewDay: { type: string, format: date }
        viewPrice: { type: number }
        direction:
          type: string
          enum: [bullish, neutral, bearish, mixed]
        contribution:
          type: number
          exclusiveMinimum: 0
        horizon: { type: string }
        thesis: { type: string }
        evidenceURL: { type: [string, "null"], format: uri }
    SmartAccountFirstOpinion:
      type: object
      description: Earliest settled Score-positive view for this platform, author and representative ticker, not a lifetime-first claim. Price is a completed daily close known before publication, never a later entry price.
      required: [publishedAt, direction, priceBasis]
      properties:
        publishedAt: { type: string, format: date-time }
        direction: { type: string, enum: [bullish, bearish] }
        price: { type: [number, "null"], exclusiveMinimum: 0 }
        priceDay: { type: [string, "null"], format: date }
        priceSource: { type: [string, "null"] }
        priceBasis: { type: string, enum: [last_completed_daily_close] }
        sourcePostId: { type: [string, "null"] }
        evidenceURL: { type: [string, "null"], format: uri }
...
    SmartAccountSettlementEvidence:
      type: object
      description: Direction-aware historical settlement against both the S&P 500 and the mapped industry ETF.
      required: [status, horizon]
      properties:
        status: { type: string, enum: [pending, settled, unavailable] }
        horizon: { type: string }
        entryDay: { type: [string, "null"], format: date }
        exitDay: { type: [string, "null"], format: date }
        entryPrice: { type: [number, "null"] }
        exitPrice: { type: [number, "null"] }
        tickerReturnPercent: { type: [number, "null"] }
        marketBenchmarkReturnPercent: { type: [number, "null"] }
        marketExcessReturnPercent: { type: [number, "null"] }
        actualHit: { type: [boolean, "null"] }
        contribution: { type: [number, "null"] }
        industryBenchmarkTicker: { type: [string, "null"] }
        industryBenchmarkReturnPercent: { type: [number, "null"] }
        industryExcessReturnPercent: { type: [number, "null"] }
        industryActualHit: { type: [boolean, "null"] }
        settlementVersion: { type: [string, "null"] }
    OpinionSupportingSource:
      type: object
      required: [id, eventId, ticker, publisher, sourceType, relationship, status, title, summary, publishedAt, sourceURL, claim, excerpt, locator]
      properties:
        id: { type: string }
        eventId: { type: string }
        ticker: { type: string }
        publisher: { type: string }
        sourceType:
          type: string
          enum: [company, regulatory, news, research, data, other]
        relationship:
          type: string
          enum: [cited, related, follow_up]
        status:
          type: string
          enum: [ready, withdrawn]
        title: { type: string }
        titleZH: { type: [string, "null"] }
        summary: { type: string }
        summaryZH: { type: [string, "null"] }
        publishedAt: { type: string, format: date-time }
        sourceURL: { type: string, format: uri, pattern: '^https://' }
        claim: { type: string }
        excerpt: { type: string }
        locator: { type: string }
        contextNote: { type: [string, "null"] }
        contextNoteZH: { type: [string, "null"] }
        updatedAt: { type: [string, "null"], format: date-time }
    SmartAccountUpdate:
      type: object
      required:
        [id, ticker, companyName, authorId, authorName, platform, score, platformPercentile, direction, lifecycle, horizon, thesis, publishedAt]
      properties:
        id: { type: string, format: uuid }
        ticker: { type: string }
        companyName: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        platform: { type: string }
        score: { type: number }
        platformPercentile:
          type: number
          minimum: 0
          maximum: 1
        direction:
          type: string
          enum: [bullish, neutral, bearish, mixed]
        lifecycle:
          type: string
          enum: [new, strengthened, weakened, reversed, closed, invalidated]
        horizon: { type: string }
        targetPrice: { type: [number, "null"] }
        thesis: { type: string }
        invalidation: { type: [string, "null"] }
        publishedAt: { type: string, format: date-time }
        sourcePostId: { type: [string, "null"] }
        sourceURL: { type: [string, "null"], format: uri }
        ingestedAt: { type: [string, "null"], format: date-time }
        processedAt: { type: [string, "null"], format: date-time }
        evidenceURL: { type: [string, "null"], format: uri }
        authorAvatarURL: { type: [string, "null"], format: uri }
        authorFollowersCount: { type: [integer, "null"], minimum: 0 }
        authorVerified: { type: [boolean, "null"] }
        originalText: { type: [string, "null"] }
        translatedText: { type: [string, "null"] }
        translatedTextZH: { type: [string, "null"] }
        translatedTextEN: { type: [string, "null"] }
        evidenceSpan: { type: [string, "null"] }
        authorScoreAsOf: { type: [string, "null"], format: date-time }
        callScoringVersion: { type: [string, "null"] }
        activityTitle: { type: [string, "null"] }
        activityTitleZH: { type: [string, "null"] }
        activityTitleEN: { type: [string, "null"] }
        supportingSources:
          description: Factual context only; absent or empty means no module. Does not affect Score.
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/OpinionSupportingSource"
        evidenceRole:
          type: [string, "null"]
          enum: [latest, strongest, counterexample, representative, null]
        representativeTickerContribution: { type: [number, "null"] }
        representativeCallCount: { type: [integer, "null"], minimum: 1 }
        representativeTickerRank: { type: [integer, "null"], minimum: 1, maximum: 3 }
        firstOpinion: { $ref: "#/components/schemas/SmartAccountFirstOpinion" }
        settlement:
          anyOf:
            - $ref: "#/components/schemas/SmartAccountSettlementEvidence"
            - type: "null"
        priceEvidence:
          anyOf:
            - $ref: "#/components/schemas/SmartAccountPriceEvidence"
            - type: "null"
...